+++ http/CHANGES.md
@@ -1,6 +1,6 @@
 # unreleased 0.6.0
 ## Change
-- bump MSRV to `1.78`
+- bump MSRV to `1.79`
 - update `xitca-io` to `0.4.0`
 - update `xitca-router` to `0.3.0`
 - update `xitca-service` to `0.2.0`