+++ http/CHANGES.md
@@ -6,6 +6,7 @@
 - update `xitca-service` to `0.2.0`
 - update `xitca-tls` to `0.4.0`
 - update `xitca-unsafe-collection` to `0.2.0`
+- update `tokio-uring` to `0.5.0`
 
 # 0.5.0
 ## Change